fix: Force pipelining for loop in propagate
hw_emu build/program was failing with Run vpl: Step config_hw_emu.elaborate: Failed
or
[libprotobuf ERROR google/protobuf/message_lite.cc:118] Can't parse message of type "response_packet_info" because it is missing required fields: size
However, it only failed when MAX_FANOUT was 64 or lesser. It seems that 64
is the cutoff for auto-pipelining which is required for a succesful build
(maybe something to do with scheduling? - builds
that would fail spent a long time in scheduling)