fix: only bump literals assigned at the conflict level
This explains the performance difference between the clause learning
branch and the main branch - the main branch is only sensitive to literals
on the trail. This behavior is probably what I wanted anyways (but should
check the Move to Front paper to be sure).

This change brings the clause learning branch back to outperforming the main
branch even on low conflict problems (multiplier_248521_sat).