feat: preemptively cancel assignment that weren't propagated
The fixes another aspect of the confusing issue of the main branch outperforming
the clause learning branch. The main (old) architecture delays assignments
via the proagation queue, so the assignments post-conflict never get
involved during conflict analysis.

This is in contrast to the clause learning branch which assigns immediately.
To make sure these "guarenteed useless assignments" don't mess up the
conflict analysis + bumping, I cancel down the the head of the propagation
queue.

It's interesting to note that CNF SAT doesn't have this issue because a
 CNF clause only can generate 1 implication. So if we resolve on a CNF clause,
 it must have been because its implication was stamped.

 This solver can generate multiple implication from the same Node, so if one of them
 is "post conflict" in the trail, it still was geting stamped and bumped.

 Now clause learning REALLY seems to beat the main branch.