feat: skip propagate if the occurrence is the cause of the assignment
If a gate caused this assignment via implication, reimplying this gate with
the new assignment will not result in any new information, so we are free
to skip for performance.

Very minimal effects on performance, there is some noise just because we
may end up implying on this gate because of another assignment, and this
early imply call changes the order.

For CPU this actually reduces imply() count by 17% and execution time
by 15% so it's not insignificant.