fix: resolving a learned assignment forces UIP to be the decision
Prior, the "UIP" we find during CA is not guarenteed to be a UIP.

When we resolve a learned clause, it might have only been asserting because
of the current decision. In this case, the "UIP" is not actually asserting
because the conflict depended on the orginal decision.

Since we don't save the learned clause, we must assume this as the worse case.

In theory this would cause some SAT instances to report UNSAT, but I
haven't actually noticed that.