feat: suppress bumping after resolving a learned clause feat: drop clause if it's too big
With this change, the clause learning branch outperforms the old branch
on the UNSAT multiplier suite.

With CL deactivated properly (using a false return if we resolve on a
"deleted" learned clause), the new architecture compares favorably to the old.

This feature suggests that the bump heuristic is VERY imporant and should
be experimented with to find a good set-up.

Dropped clauses have a constant node id as antecedent so we can detect an
attempt to resolve on a dropped clause.