fix: inheriting from ap_uint is not safe
ap_uint<18> temp = ap_uint<18>("0b011111111111111111");
    (*propagation_count) = temp;

is not the same as doing it with Pins (a struct that inherits from ap_uint<18>.
When compiled for hw_emu (but not sw_emu!), the stuct version has extra
trailing zeros that mess up the indexing. Maybe for hw_emu the ap_uint
implementation differs from a struct for memory offsets or something.

I was feeling that this wasn't safe, but Pins was the last holdover
I hadn't fixed