feat: only enqueue first new implication from gate imply
II is bounded by
1. reads from assigns
2. imply
3. writes to assigns.

By restricting the loop to enqueue JUST the first new implication, we reduce
the potential number of assigns writes from PINS_PER_GATE to 1.
This results in a smaller II. (16 to 10) (in theory it should be a
reduction of 8 ( 9 (PINS_PER_GATE) potential writes to 1). Maybe there's
some overhead we could reduce)

However, there may still be new implications on the gate that haven't
been enqueued. To cover this case, we remove the early termination check
for the case that we have already implied on this gate.
Thus, when we process the first implication, this gate will show up in its
occurences and we will get another chance to enqueue a second new implication.

I'm not even sure if that performance skip was useful in HW since
everything was pipelined anyways.