feat: reduce resolve_clause_loop latency (9 to 8)
By elimating the single use temporary clause registers, we save a cycle.
latency is the important factor because it will never be pipelined.

It seems like array_paritioning Clause.literals at any stage results in
the whole clauses array having the same partioning. For clarity I move this
to top level with
#pragma HLS array_partition variable = (*clauses).literals complete
It's not syntax, but it works: each literals index can be accessed
in parallel; and not just clauses[0] - it puts each literals[i] in a
different BRAM.