feat: burst process imply()
Profiling suggests the majority of time is still spent in Propagate.

We have been contrained by the assigns dependency between collecting
the assigns and enqueueing the new implications.
By seperating the enqueues we can pipeline the implication loop at a
much higher rate.

This may affect the order of implications in the trail, but any new
implications that would affect the gates will have their related
gates probed.