cleanup: remove location and AssertingLocation()
This was previously used (with minimal affect), to prevent resolving on
future assignments that had entered the trail and were related to the
conflict gate.

Now it makes no practical difference  (the cancelUntil II is slighly faster
than the trail pop unstamped loop, but the overhead of AsssertingLocation
made it barely worse.