feat: increase var_inc initial value
On certain problems I saw a dramatic relative increase in
conflicts/execution time (especially
../benchmarks/multiplier_factorize/multiplier_36542201_sat.eqn)
I think this is because we rescore a bit more often with deterministic
and small values are sent to 0 or lose a lot of precision. To counteract
this, I initialize the var_inc at in the middle of the FLT range instead of
FLT_MIN. This will mean an extra rescore at the start, but hopefully
reduce early precision decay.