feat: use custom data type Level for decision levels
This slightly reduces memory and cleans up the code a bit, but isn't really impactful.

Weirdly, the reported utilization is higher because the Level based
level_assigned gets explicately mapped to BRAM, but uint32_t based just has
zeros in memory columns - maybe it gets mapped later or there is a problem
with reporting URAM mapping?