feat: custom Stamp type reduce bits per stamp from 17 to 16
We now embed the unstamped id as the -1 of uint16_t instead of uint32_t.
This has higher reported utilization because Vitis explicitly puts stamps
in BRAM now.

It was cool that Vitis could detect that we were only storing uint32_t(-1)
and uint16_t in our uint32_t stamps array and trucated it accordingly.
However, it is cleaner to have it as a custom data type.